+++ tests/functional/Extension/Footnote/FootnoteExtensionTest.php
@@ -58,6 +58,26 @@ final class FootnoteExtensionTest extends TestCase
 <div class="footnotes" role="doc-endnotes"><hr /><ol><li class="footnote" id="customfn:note1" role="doc-endnote"><p>There&nbsp;<a class="footnote-backref" rev="footnote" href="#customfnref:note1" role="doc-backlink">↩</a></p></li></ol></div>',
                 ['ref_id_prefix' => 'customfnref:', 'footnote_id_prefix' => 'customfn:'],
             ],
+            [
+                // Only the first definition of a label is used, and the others are removed
+                "Here[^note1]\n\n[^note1]: First\n\n[^note1]: Second",
+                '<p>Here<sup id="fnref:note1"><a class="footnote-ref" href="#fn:note1" role="doc-noteref">1</a></sup></p>
+<div class="footnotes" role="doc-endnotes"><hr /><ol><li class="footnote" id="fn:note1" role="doc-endnote"><p>First&nbsp;<a class="footnote-backref" rev="footnote" href="#fnref:note1" role="doc-backlink">↩</a></p></li></ol></div>',
+            ],
+            [
+                // Labels are not key paths, so one label being a prefix of another is harmless
+                "[^a] and [^a.b]\n\n[^a]: note a\n\n[^a.b]: note ab",
+                '<p><sup id="fnref:a"><a class="footnote-ref" href="#fn:a" role="doc-noteref">1</a></sup> and <sup id="fnref:a.b"><a class="footnote-ref" href="#fn:a.b" role="doc-noteref">2</a></sup></p>
+<div class="footnotes" role="doc-endnotes"><hr /><ol><li class="footnote" id="fn:a" role="doc-endnote"><p>note a&nbsp;<a class="footnote-backref" rev="footnote" href="#fnref:a" role="doc-backlink">↩</a></p></li>
+<li class="footnote" id="fn:a.b" role="doc-endnote"><p>note ab&nbsp;<a class="footnote-backref" rev="footnote" href="#fnref:a.b" role="doc-backlink">↩</a></p></li></ol></div>',
+            ],
+            [
+                // "." and "/" are interchangeable key path delimiters, so these must stay distinct
+                "[^a.b] and [^a/b]\n\n[^a.b]: dot\n\n[^a/b]: slash",
+                '<p><sup id="fnref:a.b"><a class="footnote-ref" href="#fn:a.b" role="doc-noteref">1</a></sup> and <sup id="fnref:a/b"><a class="footnote-ref" href="#fn:a/b" role="doc-noteref">2</a></sup></p>
+<div class="footnotes" role="doc-endnotes"><hr /><ol><li class="footnote" id="fn:a.b" role="doc-endnote"><p>dot&nbsp;<a class="footnote-backref" rev="footnote" href="#fnref:a.b" role="doc-backlink">↩</a></p></li>
+<li class="footnote" id="fn:a/b" role="doc-endnote"><p>slash&nbsp;<a class="footnote-backref" rev="footnote" href="#fnref:a/b" role="doc-backlink">↩</a></p></li></ol></div>',
+            ],
         ];
     }